data type fixed

+++ cartel/cartel/src/main/java/com/project/cartel/dto/UserBasicDetailsDTO.java
@@ -19,7 +19,7 @@ public class UserBasicDetailsDTO {
     private Long  phone_number;
     private String email;
     private String self_referral_code;
-    private int aadhaar_number;
+    private long aadhaar_number;
     private String pan_card;
     private String upi_id;
     private Long account_no;
